pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel", "build", "uv", "pip"]
build-backend = "setuptools.build_meta"

[project]
name = "oarc-crawlers"
version = "0.1.5"
description = "OARC's dynamic webcrawler module collection"
readme = "README.md"
requires-python = ">=3.10,<3.12"
license = "Apache-2.0"
authors = [
    {name = "OARC Team"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "aiohttp>=3.11.16",
    "arxiv>=2.2.0",
    "beautifulsoup4>=4.13.3",
    "build>=1.2.2.post1",
    "click>=8.1.8",
    "duckduckgo-search>=8.0.1",
    "fastmcp>=0.2.0",
    "GitPython>=3.1.44",
    "llama-index-experimental>=0.5.5",
    "moviepy>=2.1.2",
    "mypy>=1.3.0",
    "oarc-log>=0.1.0",
    "oarc-utils>=0.1.0",
    "pandas>=2.2.3",
    "pyarrow>=19.0.1",
    "pygithub>=2.6.1",
    "pyinstaller>=6.0.0",
    "pytchat>=0.5.0",
    "pytube>=15.0.0",
    "questionary>=2.1.0",
    "twine>=5.0.0",
    "types-beautifulsoup4>=4.9.0",
    "websockets>=12.0",
]

[project.optional-dependencies]
dev = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "pytest-cov>=6.1.1",
    "ruff>=0.0.240",
]

# Define command-line scripts
[project.scripts]
oarc-crawlers = "oarc_crawlers.main:main"

[project.urls]
"Homepage" = "https://github.com/Ollama-Agent-Roll-Cage/oarc-crawlers"
"Bug Tracker" = "https://github.com/Ollama-Agent-Roll-Cage/oarc-crawlers/issues"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["oarc_crawlers", "oarc_crawlers.cli"]

[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false

[tool.pytest.ini_options]
testpaths = ["src/tests"]
python_files = "test_*.py"
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "B", "I"]
ignore = []
